+++ resteasy-spring-boot-starter/src/main/java/com/paypal/springboot/resteasy/ResteasyEmbeddedServletInitializer.java
@@ -1,13 +1,10 @@
 package com.paypal.springboot.resteasy;
 
 import org.apache.commons.io.FilenameUtils;
-import org.jboss.resteasy.core.AsynchronousDispatcher;
 import org.jboss.resteasy.plugins.servlet.ResteasyServletInitializer;
 import org.reflections.Reflections;
 import org.reflections.scanners.SubTypesScanner;
-import org.reflections.scanners.TypeAnnotationsScanner;
 import org.reflections.util.ClasspathHelper;
-import org.reflections.util.FilterBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeansException;
@@ -16,7 +13,6 @@ import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 import org.springframework.beans.factory.config.ConstructorArgumentValues;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
 import org.springframework.beans.factory.support.GenericBeanDefinition;
-import org.springframework.stereotype.Component;
 
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.Path;
@@ -24,18 +20,19 @@ import javax.ws.rs.core.Application;
 import javax.ws.rs.ext.Provider;
 import java.io.File;
 import java.net.URL;
-import java.util.*;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * This is a Spring version of {@link ResteasyServletInitializer}
  *
  * @author Fabio Carvalho (facarvalho@paypal.com or fabiocarvalho777@gmail.com)
  */
-@Component
 public class ResteasyEmbeddedServletInitializer implements BeanFactoryPostProcessor {
 
     private Set<Class<? extends Application>> applications;
-    private Set<Class<?>> resources;
+    private Set<Class<?>> allResources;
     private Set<Class<?>> providers;
 
     private static final Logger logger = LoggerFactory.getLogger(ResteasyEmbeddedServletInitializer.class);
@@ -66,11 +63,10 @@ public class ResteasyEmbeddedServletInitializer implements BeanFactoryPostProces
     }
 
     /**
-     * Scan the Classpath searching for classes of type {@link Application}, or
-     * classes annotated with {@link Path}, {@link Provider}
+     * Scan the Classpath searching for classes of type {@link Application}
      */
-    private void findJaxrsClasses() {
-        logger.debug("Finding JAX-RS classes");
+    private void findJaxrsApplicationClasses() {
+        logger.debug("Finding JAX-RS Application classes");
 
         final Collection<URL> systemPropertyURLs = ClasspathHelper.forJavaClassPath();
         final Collection<URL> classLoaderURLs = ClasspathHelper.forClassLoader();
@@ -82,16 +78,50 @@ public class ResteasyEmbeddedServletInitializer implements BeanFactoryPostProces
 
         logger.debug("Classpath URLs to be scanned: " + classpathURLs);
 
-        Reflections reflections = new Reflections(
-                classpathURLs,
-                new SubTypesScanner(),
-                new TypeAnnotationsScanner(),
-                new FilterBuilder().excludePackage(AsynchronousDispatcher.class)
-        );
+        Reflections reflections = new Reflections(classpathURLs, new SubTypesScanner());
 
         applications = reflections.getSubTypesOf(Application.class);
-        resources = reflections.getTypesAnnotatedWith(Path.class);
-        providers = reflections.getTypesAnnotatedWith(Provider.class);
+
+        if(logger.isDebugEnabled()) {
+            for (Object appClass : applications.toArray()) {
+                logger.debug("JAX-RS Application class found: {}", ((Class<Application>) appClass).getName());
+            }
+        }
+    }
+
+    /**
+     * Search for JAX-RS resource and provider Spring beans,
+     * which are the ones whose classes are annotated with
+     * {@link Path} or {@link Provider} respectively
+     *
+     * @param beanFactory
+     */
+    private void findJaxrsResourcesAndProviderClasses(ConfigurableListableBeanFactory beanFactory) {
+        logger.debug("Finding JAX-RS resources and providers Spring bean classes");
+
+        String[] resourceBeans = beanFactory.getBeanNamesForAnnotation(Path.class);
+        String[] providerBeans = beanFactory.getBeanNamesForAnnotation(Provider.class);
+
+        allResources = new HashSet<Class<?>>();
+        providers = new HashSet<Class<?>>();
+
+        for(String resourceBean : resourceBeans) {
+            allResources.add(beanFactory.getType(resourceBean));
+        }
+        for(String providerBean : providerBeans) {
+            providers.add(beanFactory.getType(providerBean));
+        }
+
+        if(logger.isDebugEnabled()) {
+            for (Object resourceClass : allResources.toArray()) {
+                logger.debug("JAX-RS resource class found: {}", ((Class) resourceClass).getName());
+            }
+        }
+        if(logger.isDebugEnabled()) {
+            for (Object providerClass: providers.toArray()) {
+                logger.debug("JAX-RS provider class found: {}", ((Class) providerClass).getName());
+            }
+        }
     }
 
     /**
@@ -103,7 +133,7 @@ public class ResteasyEmbeddedServletInitializer implements BeanFactoryPostProces
         if (applications != null && applications.size() != 0) {
             return false;
         }
-        if (resources != null && resources.size() != 0) {
+        if (allResources != null && allResources.size() != 0) {
             return false;
         }
         if (providers != null && providers.size() != 0) {
@@ -116,7 +146,8 @@ public class ResteasyEmbeddedServletInitializer implements BeanFactoryPostProces
     public void postProcessBeanFactory(ConfigurableListableBeanFactory beanFactory) throws BeansException {
         logger.debug("Post process bean factory has been called");
 
-        findJaxrsClasses();
+        findJaxrsApplicationClasses();
+        findJaxrsResourcesAndProviderClasses(beanFactory);
 
         if (noClasses()) {
             logger.warn("No JAX-RS classes have been found");
@@ -150,6 +181,8 @@ public class ResteasyEmbeddedServletInitializer implements BeanFactoryPostProces
         GenericBeanDefinition applicationServletBean = new GenericBeanDefinition();
         applicationServletBean.setFactoryBeanName(ResteasyApplicationBuilder.BEAN_NAME);
         applicationServletBean.setFactoryMethodName("build");
+
+        Set<Class<?>> resources = allResources;
 
         ConstructorArgumentValues values = new ConstructorArgumentValues();
         values.addIndexedArgumentValue(0, applicationClass.getName());